Remove accessToken from PDF URLs
Access tokens and bearer tokens should not be passed as part of the
query string. Per the [OAuth RFC]:

> Don't pass bearer tokens in page URLs:  Bearer tokens SHOULD NOT be
> passed in page URLs (for example as query string parameters).
> Instead, bearer tokens SHOULD be passed in HTTP message headers or
> message bodies for which confidentiality measures are taken.
> Browsers, web servers, and other software may not adequately
> secure URLs in the browser history, web server logs, and other
> data structures.  If bearer tokens are passed in page URLs,
> attackers might be able to steal them from the history data, logs,
> or other unsecured locations.

This replaces the query parameter with a header. This means that
standard HTTP(S) URLs won't work, and will need to be fetched by the
app. That limitation could be fixed with cookies, but that takes extra
care and shouldn't be done trivially.